Implementacion de memory cache para operaciones de listado de productos

+++ Infrastructure/Data/Repositories/ProductRepository.cs
@@ -11,12 +11,18 @@ internal class ProductRepository : IProductRepository
     private readonly ECommerceContext _context;
 
     /// <summary>
+    /// 
+    /// </summary>
+    private readonly ICachingManager _cachingManager;
+
+    /// <summary>
     ///
     /// </summary>
     /// <param name="context"></param>
-    public ProductRepository(ECommerceContext context)
+    public ProductRepository(ECommerceContext context, ICachingManager cachingManager)
     {
         _context = context;
+        _cachingManager = cachingManager;
     }
 
     /// <summary>
@@ -37,6 +43,10 @@ internal class ProductRepository : IProductRepository
             _context.Products.Add(product);
         }
         await _context.SaveChangesAsync();
+
+        // Borrar el cache de productos para obligar a actualizarla de nuevo
+        _cachingManager.Clear(CachingManager.ProductList);
+
         return product;
     }
 
@@ -51,9 +61,18 @@ internal class ProductRepository : IProductRepository
 
         if (product is not null)
         {
-            _context.Products.Remove(product);
-            await _context.SaveChangesAsync();
-            return true;
+            bool productContainsOrders = _context.OrderItems.Any(x => x.ProductId == productId);
+
+            if (!productContainsOrders)
+            {
+                _context.Products.Remove(product);
+                await _context.SaveChangesAsync();
+
+                // Borrar el cache de productos para obligar a actualizarla de nuevo
+                _cachingManager.Clear(CachingManager.ProductList);
+                return true;
+            }
+            else throw new Exception("el producto tiene ordenes asociadas");           
         }
 
         return false;
